refactoring data for step challenge - moving logic functions to separate JS file
